First pass of picking up objects
Can now use the B button (stacked with the stealth move for now, change as needed) to pick up any item that has a holdable script. Player will need to have a Holder Script. This allow the player to pick up and put down items and the target, and the holdable script is already set up to be paired with a moving object (the target), while also working with stationary items. You can also edit the displacement the obejct is away from you when you put it on the ground again.