Guard AI now detects player only when in visibility cone.
Guard AI will now only detect the player when they are within range and in front of the guard within a cone of a certain angle. Cone detection uses dot product to solve for detection angle.  Angle is default set to 45 degrees but can be adjusted through inspector. Sound detection coming soon.