Guard Scanning behaviour now implemented
Guards can now scan the environment for the player, which means rotating a certain degree to the left and right to be on the lookout for players. The angle of the scan can be specified, as well as whether guards will scan at each pathpoint they go to. Scanning also occurs if a guard is suspicious and reaches the source of a noise. Scanning can also be set to specific directions by changing the direction of the path points, which will be the general direction that the guard's scan takes place when they get there.