Gadgets Implementation
The Cake, Popsicle and Jazz gadgets are in (Popsicle and Jazz are shorthand for the full names). All follow the specifics laid out in JIRA: the cake can be placed and, if seen by an enemy will distract them as they move toward and eat it. Jazz will immediatly distract those in the vicinity as well as fisabling sound detectino while they are distracted. The Popsicle distracts as well, but still allows for visual and sound detection. The popsickle can also be "slid" across the ground as whenever you place it your speed if factored into the speed the popsicle goes, allowing you to throw it towards guards in certain directions. These gadgets come with various tweakable parameters like range and duration time.