Added line-of-sight guard AI systems, minor fixes
Guards now have a "Player Out of Sight Timer" which is the amount of time a guard will continue a chase when a player has gone out of sight. Also finally pushing some extra functionality for the target end portal (to make it actually work), and clamping the guard flashlight angle.